+++ src/pages/settings/company-import-export/components/Backup.tsx
@@ -8,14 +8,40 @@
  * @license https://www.elastic.co/licensing/elastic-license
  */
 
-import { UploadImport } from 'components/import/UploadImport';
+import { Card, Element } from '@invoiceninja/cards';
+import { Button } from '@invoiceninja/forms';
+import { endpoint } from 'common/helpers';
+import { request } from 'common/helpers/request';
+import { toast } from 'common/helpers/toast/toast';
+import { FormEvent } from 'react';
+import { useTranslation } from 'react-i18next';
 
 export function Backup() {
+  const [t] = useTranslation();
+
+  const handleExportCompany = (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+
+    toast.processing();
+
+    request('POST', endpoint('/api/v1/export'), {
+      send_email: true,
+      report_keys: [],
+    })
+      .then((response) => {
+        toast.success(response.data.message);
+      })
+      .catch((error) => {
+        console.error(error);
+        toast.error();
+      });
+  };
+
   return (
-    <div className="grid grid-cols-12">
-      <div className="col-span-12">
-        <UploadImport entity="company" onSuccess={false} type="zip" />
-      </div>
-    </div>
+    <Card>
+      <Element leftSide={t('export_company')}>
+        <Button onClick={handleExportCompany}>{t('export')}</Button>
+      </Element>
+    </Card>
   );
 }
