+++ src/pages/settings/routes.tsx
@@ -52,12 +52,9 @@ export const settingsRoutes = (
     <Route path="workflow_settings" element={<Settings.WorkflowSettings />} />
     <Route path="import_export" element={<Settings.ImportExport />} />
     <Route path="account_management" element={<Settings.AccountManagement />} />
-    <Route
-      path="company_import_export"
-      element={<Settings.CompanyImportExport />}
-    >
+    <Route path="backup_restore" element={<Settings.CompanyBackupRestore />}>
       <Route path="" element={<Settings.CompanyBackup />} />
-      <Route path="export" element={<Settings.CompanyExport />} />
+      <Route path="restore" element={<Settings.CompanyRestore />} />
     </Route>
     <Route
       path="invoice_design"
